Add an option whether to check updates or not for pyg users

+++ ogb/graphproppred/dataset_pyg.py
@@ -9,7 +9,7 @@ from ogb.io.read_graph_pyg import read_graph_pyg
 
 
 class PygGraphPropPredDataset(InMemoryDataset):
-    def __init__(self, name, root = 'dataset', transform=None, pre_transform = None, meta_dict = None):
+    def __init__(self, name, root = 'dataset', transform=None, pre_transform = None, meta_dict = None, check_update = True):
         '''
             - name (str): name of the dataset
             - root (str): root directory to store the dataset folder
@@ -50,7 +50,7 @@ class PygGraphPropPredDataset(InMemoryDataset):
         # First check whether the dataset has been already downloaded or not.
         # If so, check whether the dataset version is the newest or not.
         # If the dataset is not the newest version, notify this to the user. 
-        if osp.isdir(self.root) and (not osp.exists(osp.join(self.root, 'RELEASE_v' + str(self.meta_info['version']) + '.txt'))):
+        if check_update and osp.isdir(self.root) and (not osp.exists(osp.join(self.root, 'RELEASE_v' + str(self.meta_info['version']) + '.txt'))):
             print(self.name + ' has been updated.')
             if input('Will you update the dataset now? (y/N)\n').lower() == 'y':
                 shutil.rmtree(self.root)
